navigation improvements
